backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
gunicorn==25.3.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
pydantic[email]==2.12.5
python-dotenv==1.2.2
python-multipart==0.0.22
slowapi==0.1.9
openpyxl==3.1.5
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
google-genai==1.68.0
pytesseract==0.3.13
pymupdf==1.27.2.2
pillow==12.1.1
email-validator==2.3.0
cachetools==7.0.5
structlog==25.5.0
